internal: move generated files into internal/api
