+++ cmd/flightctl-server/main.go
@@ -12,12 +12,12 @@ import (
 	cacheutil "github.com/argoproj/argo-cd/v2/util/cache"
 	oapimiddleware "github.com/deepmap/oapi-codegen/pkg/chi-middleware"
 	api "github.com/flightctl/flightctl/api/v1alpha1"
+	"github.com/flightctl/flightctl/internal/api/server"
 	"github.com/flightctl/flightctl/internal/config"
 	"github.com/flightctl/flightctl/internal/configprovider/git"
 	"github.com/flightctl/flightctl/internal/crypto"
 	device_updater "github.com/flightctl/flightctl/internal/monitors/device-updater"
 	"github.com/flightctl/flightctl/internal/monitors/repotester"
-	"github.com/flightctl/flightctl/internal/server"
 	"github.com/flightctl/flightctl/internal/service"
 	"github.com/flightctl/flightctl/internal/store"
 	"github.com/flightctl/flightctl/pkg/log"
