+++ cmd/flightctl/main.go
@@ -17,7 +17,7 @@ import (
 	"text/tabwriter"
 
 	api "github.com/flightctl/flightctl/api/v1alpha1"
-	"github.com/flightctl/flightctl/internal/client"
+	"github.com/flightctl/flightctl/internal/api/client"
 	"github.com/flightctl/flightctl/internal/config"
 	"github.com/flightctl/flightctl/internal/crypto"
 	"github.com/spf13/cobra"
