+++ internal/agent/enrollment.go
@@ -13,7 +13,7 @@ import (
 	"time"
 
 	api "github.com/flightctl/flightctl/api/v1alpha1"
-	"github.com/flightctl/flightctl/internal/client"
+	"github.com/flightctl/flightctl/internal/api/client"
 	fccrypto "github.com/flightctl/flightctl/internal/crypto"
 	"github.com/skip2/go-qrcode"
 	"k8s.io/apimachinery/pkg/util/wait"
