+++ internal/api/server/docs.go
@@ -1,0 +1,3 @@
+package server
+
+//go:generate go run github.com/deepmap/oapi-codegen/cmd/oapi-codegen@v1.15.0 --config=server.gen.cfg ../../../api/v1alpha1/openapi.yaml
