--- internal/client/docs.go
@@ -1,3 +1,0 @@
-package client
-
-//go:generate go run github.com/deepmap/oapi-codegen/cmd/oapi-codegen@v1.15.0 --config=client.gen.cfg ../../api/v1alpha1/openapi.yaml
