--- internal/server/docs.go
@@ -1,3 +1,0 @@
-package server
-
-//go:generate go run github.com/deepmap/oapi-codegen/cmd/oapi-codegen@v1.15.0 --config=server.gen.cfg ../../api/v1alpha1/openapi.yaml
